Update ode utils (#94)
* Update ode_utils.py

merge ode_solver into ide_utils

* Delete ngclearn/utils/diffeq/ode_solver.py

* Update ode_utils.py

refactor doc-string
